+++ workmanager/test/workmanager_test.mocks.dart
@@ -35,12 +35,14 @@ class MockWorkmanager extends _i1.Mock implements _i2.Workmanager {
 
   @override
   _i3.Future<void> initialize(
-    Function? callbackDispatcher,
-  ) =>
+    Function? callbackDispatcher, {
+    bool? isInDebugMode = false,
+  }) =>
       (super.noSuchMethod(
         Invocation.method(
           #initialize,
           [callbackDispatcher],
+          {#isInDebugMode: isInDebugMode},
         ),
         returnValue: _i3.Future<void>.value(),
         returnValueForMissingStub: _i3.Future<void>.value(),